refactor: 클래스 이름 수정
프로그램에 대한 예외 상황 문구를 관리하는 클래스로 수정하기 위해 클래스 이름을 수정하였습니다.
