+++ src/main/java/lotto/model/LottoNumber.java
@@ -3,7 +3,7 @@ package lotto.model;
 
 import static lotto.model.LottoAttribute.MAX_LOTTO_NUMBER;
 import static lotto.model.LottoAttribute.MIN_LOTTO_NUMBER;
-import static lotto.model.LottoErrorType.*;
+import static lotto.model.ErrorType.*;
 
 import java.util.Objects;
